dodo

+++ DjangoWithReact/frontend/gui/src/containers/ArticleDetailView.js
@@ -1,7 +1,6 @@
 import React from 'react';
 import axios from 'axios'
-
-
+import {connect} from 'react-redux'
 import {Card,Button} from 'antd'
 import CustomForm from  '../components/Form'
 
@@ -12,7 +11,14 @@ class ArticleDetail extends React.Component {
         article : {}
     }
 
-    componentDidMount () {
+
+    componentWillReceiveProps (newProps) {
+        console.log(newProps)
+        if(newProps.token){
+            axios.defaults.headers = {
+                "Content-Type" : "application/json",
+                Authorization : newProps.token
+            }
             const articleID = this.props.match.params.articleID
             axios.get(`http://127.0.0.1:8000/api/${articleID}/`)
             .then (res => {
@@ -21,12 +27,24 @@ class ArticleDetail extends React.Component {
                 });
 
             })
+        }
+       
     }
+
+  
     handleDelete =  (event) => {
+        if(this.props.token !== null) {
+            axios.defaults.headers = {
+                "Content-Type" : "application/json",
+                Authorization : this.props.token
+            }
             const articleID = this.props.match.params.articleID
             axios.delete(`http://127.0.0.1:8000/api/${articleID}/`)
             this.props.history.push('/');
             this.forceUpdate();
+        }else {
+            //show some message
+        }
 
     }
     render () {
@@ -43,4 +61,10 @@ class ArticleDetail extends React.Component {
         )
     }
 } 
-export default ArticleDetail;
+const  mapStateTpProps = state => {
+    return {
+      token : state.token 
+    } 
+}
+
+export default connect(mapStateTpProps)(ArticleDetail);
